complete auth v3
edit AuthFilter
